sample result
due to the "bug" that i left on the first script, i couldnt download all the issue page, so i did an test using some of the pages that i was able to download (3,4,6,7,9,10,11,12,13,14) to use as sample base.
i also manually filtered the result to remove any noise.
judging by the number of lines, there are 228 games/softwares on this sample base, if we consider that we only downloaded 10/125 pages, the correct number should be something like 2918 games/softwares, feel free to download all the pages to see if the numbers are any close to that. (i will do that my self, as soon as github allow me to download more stuff)
keep in mind that some games may never have being tested, so there is no issue associated with then, despite they not working correctly.